Further clean up YSO notebook, apply units consistenly
Astropy units were used in this notebook in some parts, but not in others,
which led to some inconsistencies. During the flux calibration, the
background subtraction was scaled with the exposure time of the standard
star, which I believe should not be the case. This also leads to a more
realistic scaling factor in the final spectrum comparison (unless I'm
missing something, scaling the source *down* to the result makes more
sense, as the flux should decrease as it passes through the optical train).

Also made some plots look a little nicer and simplified some of the code.