Fix test class that should be a function
A very similar thing occurred recently but I cannot find it right now: The code for this test was in the class body (as opposed to within a method), so it was executed during collection time and not a an actual test. Simple solution: just make it a function, it's just a single test case anyway.