Restructure CI workflows
The current setup is very clumsy and hacky, and in several details doesn't
do what we need, e.g. running notebook tests from "sister branches". I also
ran into the situation many times that I wanted to run them also for a PR
or just for a change in ScopeSim (without a corresponding PR in the IRDB),
to see if I'm about to break anything.

For all those reasons, I "pulled apart" that one large workflow file into
multiple separate, clean and simple workflows that rely on each other.
I also added an attempt of an explanation in the form of `CI_strategy.md`,
which should hopefully (togehter with comments in the workflow files)
serve as documentation for all of this, as well as a place to share ideas
for further improvement, as some aspects of all of this are not yet final.
Especially those badge reports are still not really used, just created.
I do have a few things in mind how to do that, but that'll be for later.