Remove all remaining `version.yaml` files
Those had been obsolete for more than one year, since we updated our
release workflow. When a package is compliled to be uploaded to the server,
a `version.yaml` file is created on the fly from the current timestamp and
packaged in the zip file, so a package installed from the server has the
version information corresponding to a tag.

But these newly created `version.yaml` files are not committed back into
the repo (on purpose), so the old ones that were still there showed wildly
outdated timestamps, which greatly confuses some users (see e.g. the post-
merge comment in #241, although that shouldn't be _there_ but whatever)
when installing from the repo.

I'm working on adding the funcitonality to `sim.bug_report()` to display a
meaningful version information when a package was installed from source,
but that will also be easier with those outdated `version.yaml` files gone.

IIRC we actually decided to remove them back when we updated the release
workflow, but then forgot about it...