Add `ExposureOutput` and `ADConversion` to MICADO
Keeping the mode for `ExposureOutput` as "sum" for now to hopefully not
break the notebooks. This can be changed later.

How could the `ADConversion` get the gain value from
`FPA_metis_img_lm_layout.dat`?

Also, some indentation fixing and spelling (When was the last time
someone looked at this file? What's with all the commented-out stuff?).