Remove redundant !!float in yamls, fix others
Since forever, anything that has a decimal point and an explicit plus or minus in the exponent is converted to a float anyway, no need for this ugly tag.
Also, "0." is converted to `0.0` anyway.

While in the process, I standardized on a lowercase "e". Both works fine and we had both before, but why not.